[refactor] Modularize CLI entry point
